services/kafka-consumers/db-writer/requirements.txt
kafka-python>=2.0.2
psycopg[binary]>=3.1.18
